+++ SOTFEdit/Model/SaveData/Actor/Influence.cs
@@ -4,9 +4,9 @@ namespace SOTFEdit.Model.SaveData.Actor;
 
 public partial class Influence : ObservableObject
 {
-    [ObservableProperty] private float _anger;
-    [ObservableProperty] private float _fear;
-    [ObservableProperty] private float _sentiment;
+    [ObservableProperty] private float? _anger;
+    [ObservableProperty] private float? _fear;
+    [ObservableProperty] private float? _sentiment;
     public string TypeId { get; init; }
 
     public static class Type